+++ api/pulumi/prod/elasticSearch.ts
@@ -53,6 +53,11 @@ class ElasticSearch {
                                 AWS: currentCallerIdentity.accountId
                             },
                             Action: "es:*",
+                            /**
+                             * Error that started with new pulumi version.
+                             * TODO: remove ts-ignore when its fixed
+                             */
+                            // @ts-ignore
                             Resource: this.domain.arn.apply(v => `${v}/*`)
                         }
                     ]