+++ packages/api-headless-cms/src/content/plugins/crud/contentEntry.crud.ts
@@ -45,6 +45,7 @@ import {
     entryFromStorageTransform,
     entryToStorageTransform
 } from "~/content/plugins/utils/entryStorage";
+import { assignAfterEntryDelete } from "~/content/plugins/crud/contentEntry/afterDelete";
 
 export const STATUS_DRAFT = "draft";
 export const STATUS_PUBLISHED = "published";
@@ -169,6 +170,10 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
         context,
         onBeforeUpdate
     });
+    assignAfterEntryDelete({
+        context,
+        onAfterDelete
+    });
 
     const checkEntryPermissions = (check: {
         rwd?: string;
@@ -215,8 +220,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
         await utils.checkModelAccess(context, model);
 
         const entries = await storageOperations.entries.getByIds(model, {
-            tenant: getTenant().id,
-            locale: getLocale().code,
             ids
         });
 
@@ -273,8 +276,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             await utils.checkModelAccess(context, model);
 
             const entries = await storageOperations.entries.getPublishedByIds(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 ids
             });
 
@@ -288,8 +289,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             await utils.checkModelAccess(context, model);
 
             const entries = await storageOperations.entries.getLatestByIds(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 ids
             });
 
@@ -298,8 +297,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
 
         getEntryRevisions: async (model, entryId) => {
             return storageOperations.entries.getRevisions(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id: entryId
             });
         },
@@ -480,15 +477,11 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             const [uniqueId] = sourceId.split("#");
 
             const originalStorageEntry = await storageOperations.entries.getRevisionById(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id: sourceId
             });
             const latestStorageEntry = await storageOperations.entries.getLatestRevisionByEntryId(
                 model,
                 {
-                    tenant: getTenant().id,
-                    locale: getLocale().code,
                     id: uniqueId
                 }
             );
@@ -597,8 +590,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
              * The entry we are going to update.
              */
             const originalStorageEntry = await storageOperations.entries.getRevisionById(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id
             });
 
@@ -688,23 +679,17 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             const { id: entryId, version } = parseIdentifier(revisionId);
 
             const storageEntryToDelete = await storageOperations.entries.getRevisionById(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id: revisionId
             });
             const latestStorageEntry = await storageOperations.entries.getLatestRevisionByEntryId(
                 model,
                 {
-                    tenant: getTenant().id,
-                    locale: getLocale().code,
                     id: entryId
                 }
             );
             const previousStorageEntry = await storageOperations.entries.getPreviousRevision(
                 model,
                 {
-                    tenant: getTenant().id,
-                    locale: getLocale().code,
                     entryId,
                     version
                 }
@@ -780,8 +765,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             await utils.checkModelAccess(context, model);
 
             const storageEntry = await storageOperations.entries.getLatestRevisionByEntryId(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id: entryId
             });
 
@@ -803,12 +786,7 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             const permission = await checkEntryPermissions({ pw: "p" });
             await utils.checkModelAccess(context, model);
 
-            const tenant = getTenant().id;
-            const locale = getLocale().code;
-
             const originalStorageEntry = await storageOperations.entries.getRevisionById(model, {
-                tenant,
-                locale,
                 id
             });
 
@@ -875,8 +853,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             const permission = await checkEntryPermissions({ pw: "c" });
 
             const originalStorageEntry = await storageOperations.entries.getRevisionById(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id
             });
 
@@ -952,15 +928,11 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
             const [entryId] = id.split("#");
 
             const originalStorageEntry = await storageOperations.entries.getRevisionById(model, {
-                tenant: getTenant().id,
-                locale: getLocale().code,
                 id
             });
             const latestEntryRevision = await storageOperations.entries.getLatestRevisionByEntryId(
                 model,
                 {
-                    tenant: getTenant().id,
-                    locale: getLocale().code,
                     id: entryId
                 }
             );
@@ -1037,8 +1009,6 @@ export const createContentEntryCrud = (params: Params): CmsEntryContext => {
 
             const originalStorageEntry =
                 await storageOperations.entries.getPublishedRevisionByEntryId(model, {
-                    tenant: getTenant().id,
-                    locale: getLocale().code,
                     id: entryId
                 });