+++ packages/api-headless-cms/src/plugins/crud/index.ts
@@ -3,6 +3,10 @@ import { createSystemCrud } from "./system.crud";
 import { CmsContext, HeadlessCmsStorageOperations } from "~/types";
 import { ContextPlugin } from "@webiny/handler/plugins/ContextPlugin";
 import { createModelGroupsCrud } from "~/content/plugins/crud/contentModelGroup.crud";
+import { createModelsCrud } from "~/content/plugins/crud/contentModel.crud";
+import { createContentEntryCrud } from "~/content/plugins/crud/contentEntry.crud";
+
+const debug = process.env.DEBUG === "true";
 
 export interface Params {
     storageOperations: HeadlessCmsStorageOperations;
@@ -16,9 +20,10 @@ export const createAdminCruds = (params: Params) => {
          * It is to make sure that we load setup context before the CRUD init in our internal code.
          */
         if (!context.cms) {
-            console.log(
-                `Missing initial "cms" on the context. Make sure that you set it up before creating Admin CRUDs.`
-            );
+            debug &&
+                console.log(
+                    `Missing initial "cms" on the context. Make sure that you set it up before creating Admin CRUDs.`
+                );
             return;
         }
         const getLocale = () => {
@@ -39,6 +44,7 @@ export const createAdminCruds = (params: Params) => {
 
         context.cms = {
             ...context.cms,
+            storageOperations,
             ...createSystemCrud({
                 context,
                 getTenant,
@@ -52,6 +58,20 @@ export const createAdminCruds = (params: Params) => {
                 storageOperations
             }),
             ...createModelGroupsCrud({
+                context,
+                getTenant,
+                getLocale,
+                getIdentity,
+                storageOperations
+            }),
+            ...createModelsCrud({
+                context,
+                getTenant,
+                getLocale,
+                getIdentity,
+                storageOperations
+            }),
+            ...createContentEntryCrud({
                 context,
                 getTenant,
                 getLocale,