+++ packages/app-headless-cms/src/admin/plugins/fields/ref.tsx
@@ -19,7 +19,8 @@ const plugin: CmsEditorFieldTypePlugin = {
     name: "cms-editor-field-type-ref",
     field: {
         type: "ref",
-        validators: [],
+        validators: ["required"],
+        listValidators: ["minLength", "maxLength"],
         label: t`Reference`,
         description: t`Reference existing content entries. For example, a book can reference one or more authors.`,
         icon: <RefIcon />,