+++ packages/app-page-builder/src/editor/plugins/elements/button/index.tsx
@@ -19,6 +19,7 @@ const buttonWrapper = css({
 const buttonElementPluginsFactory = (args: PbEditorElementPluginArgs = {}) => {
     const defaultSettings = [
         "pb-editor-page-element-style-settings-button",
+        "pb-editor-page-element-style-settings-action",
         "pb-editor-page-element-style-settings-link",
         "pb-editor-page-element-style-settings-horizontal-align-flex",
         "pb-editor-page-element-style-settings-margin",